backend/requirements.txt
# Django core.

Django>=5.1.7
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.3.0
django-cors-headers>=4.3.1


mysqlclient>=2.2.1
python-dotenv>=1.0.0
Pillow>=10.1.0
drf-yasg>=1.21.7
django-cleanup>=8.0.0
pytest-django>=4.7.0
coverage>=7.4.1
gunicorn>=21.2.0